:thumbsup: 表示系をprint_issues.pyに移動
